requirements.txt
Flask==3.0.2
gunicorn==21.2.0
pdfplumber==0.11.5
python-docx==1.1.2
spacy==3.8.3
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl
requests==2.31.0
